added support for flags of save, multithreading, and anti-aliasing

+++ src/view/camera_op.c
@@ -6,7 +6,7 @@
 /*   By: mdomnik <mdomnik@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/19 17:07:08 by mdomnik           #+#    #+#             */
-/*   Updated: 2025/02/22 20:48:44 by mdomnik          ###   ########.fr       */
+/*   Updated: 2025/03/10 14:39:19 by mdomnik          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ t_camera	*camera_new(int hsize, int vsize, float fov)
 	return (camera);
 }
 
-void	ray_for_pixel(t_camera *camera, int px, int py, t_ray *ray)
+void	ray_for_pixel(t_camera *camera, float px, float py, t_ray *ray)
 {
 	float		xoffset;
 	float		yoffset;
@@ -60,6 +60,7 @@ void	ray_for_pixel(t_camera *camera, int px, int py, t_ray *ray)
 	world_x = camera->half_width - xoffset;
 	world_y = camera->half_height - yoffset;
 	ray->orig = multiply_matrix_tuple(inv_transform, new_point3(0, 0, 0));
+	// printf("hi\n");
 	ray->dir = multiply_matrix_tuple(inv_transform, new_point3(world_x,
 				world_y, -1));
 	ray->dir = sub_tuple(ray->dir, ray->orig);
